v4/go: fail closed without a worker build and repair the cross-platform build matrix
Faultable validation, candidate inspection, and recovery now refuse
with ErrorOSUnsupported before any source scan on every platform
without a worker build (routing_other.go), matching binary-format-v4.md
section 19 and the Rust client failure when the executable is absent;
the in-process machines stay the worker-side engines and are no longer
reachable from the public routing facade. The worker package and
binary are build-tagged to the worker cross-build matrix (linux,
darwin, freebsd, windows x amd64, arm64), so off-matrix targets
compile with the package stub and the refusal entry instead of failing
on missing mapped-control assembly. The retained-directory scan decodes
dirent fields in native byte order (fixes big-endian s390x/ppc64/mips64
listing, matching Go's os package), the darwin ACL sentinels use
portable word-size expressions, and test fixtures bound page-space
checks with uint64. The Go module now cross-compiles and its test
sources vet-clean on s390x, riscv64, ppc64le, mips64, 386, and
freebsd/riscv64; the previously failing big-endian CI build is
repairable. Matrix-platform tests that exercise the worker skip on
off-matrix platforms, and a new test pins the fail-closed class on
those platforms. doc.go and the parity records state the fail-closed
stance.

+++ v4/go/internal/bitmap/free_test.go
@@ -68,7 +68,7 @@ func (m *bitmapMemoryStore) DiscardPrivate(pageNumber uint32) error {
 }
 
 func (m *bitmapMemoryStore) AllocateBitmapPage() (uint32, error) {
-	if len(m.pages) >= 1<<32 {
+	if uint64(len(m.pages)) >= 1<<32 {
 		return 0, invalid("test page space exhausted")
 	}
 	pageNumber := uint32(len(m.pages))
