+++ v4/go/internal/bitmap/used_test.go
@@ -49,7 +49,7 @@ func (m *usedMemoryStore) RestoreDirty(pageNumber uint32, tag uint32) error {
 }
 
 func (m *usedMemoryStore) Allocate() (uint32, error) {
-	if len(m.pages) >= 1<<32 {
+	if uint64(len(m.pages)) >= 1<<32 {
 		return 0, invalid("test page space exhausted")
 	}
 	pageNumber := uint32(len(m.pages))
